requirements.txt
tk
Pillow
torch
screeninfo
pyautogui
stable-baselines3
python-dotenv
transforms
torchvision